+++ lib/ice_kickstarter/rake/integration_task.rb
@@ -46,6 +46,7 @@ module IceKickstarter
       def call_generators
         Bundler.with_clean_env do
           sh "cd #{app_path} && bundle exec rails generate cms:model news"
+          sh "cd #{app_path} && bundle exec rails generate cms:component:airbrake"
           sh "cd #{app_path} && bundle exec rails generate cms:component:google_analytics"
           sh "cd #{app_path} && bundle exec rails generate cms:component:contact_page"
           sh "cd #{app_path} && bundle exec rails generate cms:component:language_switch"